+++ tests/push4-create-package.test.ts
@@ -1,0 +1,21 @@
+import { test, expect } from "bun:test"
+import { getCliTestFixture } from "./fixtures/get-cli-test-fixture"
+import * as fs from "node:fs"
+import * as path from "node:path"
+
+test("should create package if it does not exist", async () => {
+  const { tmpDir, runCommand } = await getCliTestFixture({
+    loggedIn: true,
+  })
+
+  const snippetFilePath = path.resolve(tmpDir, "snippet.tsx")
+
+  fs.writeFileSync(snippetFilePath, "// Snippet content")
+  fs.writeFileSync(
+    path.resolve(tmpDir, "package.json"),
+    JSON.stringify({ name: "test-package", version: "1.0.0" }),
+  )
+
+  const { stdout } = await runCommand(`tsci push ${snippetFilePath}`)
+  expect(stdout).toContain("Successfully pushed package")
+})
